Add half-step place ratings
Supports 0.5 saved-place rating increments across the app and preserves half-step values through local persistence, remote DTOs, and recommendation averages. Adds the Supabase rating_score numeric(2,1) migration/RPC refresh and regression coverage for half-step storage/rejection.

+++ Wander/Models/LocalListModels.swift
@@ -180,7 +180,7 @@ struct ListSuggestionPlacePayload: Codable, Equatable {
     let locality: String?
     let region: String?
     let status: PlaceStatus
-    let ratingScore: Int?
+    let ratingScore: Double?
     let recommendedScore: Double?
     let recommendedCount: Int
     let attributesText: String
